+++ crates/app/src/watchers/workspace_watcher.rs
@@ -42,6 +42,10 @@ impl WorkspaceWatcher {
 
 #[async_trait]
 impl FileWatcher<AtomicDaemonState> for WorkspaceWatcher {
+    async fn on_init(&mut self, _state: AtomicDaemonState) -> miette::Result<()> {
+        Ok(())
+    }
+
     async fn on_file_event(
         &mut self,
         state: AtomicDaemonState,
@@ -51,8 +55,8 @@ impl FileWatcher<AtomicDaemonState> for WorkspaceWatcher {
             return Ok(());
         }
 
-        // Handle root `.prototools` changes
-        if event.path.as_str() == ".prototools" {
+        // Handle `.prototools` changes
+        if event.path.ends_with(".prototools") {
             self.reset_proto(&state).await?;
 
             return Ok(());